requirements.txt
pytorch
torchvision
cudatoolkit
cudnn
matplotlib
pytesseract
termcolor
tqdm
scikit-learn
scikit-image
speechrecognition
pydub
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pandas
spacy
pillow
sox
pyaudio
isodate
appdirs
gdown
imagehash
imageio
sphinx
pre-commit
deepspeech-gpu
webrtcvad
halo
webvtt-py
efficientnet-pytorch
opencv-python
summa
sumy
symspellpy
sentence-transformers
transformers
tensorflow==2.2.0
deepsegment
pygifsicle
pytorch_lightning
jiwer
youtube_dl
imutils
opencv-contrib-python
pytorch-ranger
vosk
